Passions -> Google Image
In order to maintain the same aesthetic between our pets and passions, gets the first google images result for each passion.

Note, a lot of these are pretty small and we might have to adjust the scrape to take something later than the first result or only accept a larger image.